% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{apply_snv}
\alias{apply_snv}
\title{Apply Standard Normal Variate (SNV) row-wise}
\usage{
apply_snv(M)
}
\arguments{
\item{M}{Numeric matrix}
}
\value{
SNV-transformed matrix
}
\description{
Apply Standard Normal Variate (SNV) row-wise
}
