% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_model_cache}
\alias{clear_model_cache}
\title{Clear the in-memory model cache}
\usage{
clear_model_cache()
}
\value{
Invisible \code{NULL}
}
\description{
Releases all cached keras model objects. Models will be reloaded from disk
on the next prediction call.
}
