% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_create}
\alias{dir_create}
\title{Create a directory if it does not already exist}
\usage{
dir_create(path)
}
\arguments{
\item{path}{Directory path to create}
}
\description{
Create a directory if it does not already exist
}
