% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{family_matches}
\alias{family_matches}
\title{Test whether a family matches a given sensor and moisture combination}
\usage{
family_matches(fam, sensor, moisture)
}
\arguments{
\item{fam}{Family list from model_registry}

\item{sensor}{Character string for the sensor name}

\item{moisture}{Character string for the moisture mode}
}
\value{
Logical
}
\description{
Test whether a family matches a given sensor and moisture combination
}
