% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_family}
\alias{get_family}
\title{Lookup a model family from the registry}
\usage{
get_family(id)
}
\arguments{
\item{id}{Family ID string: \code{"OSSL_VisNIR"} or \code{"OSSL_MIR"}}
}
\value{
Named list describing the family
}
\description{
Lookup a model family from the registry
}
