% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{model_registry}
\alias{model_registry}
\title{Official autoSpectra model family registry}
\format{
An object of class \code{list} of length 2.
}
\usage{
model_registry
}
\description{
Two sensor-agnostic families trained on the full OSSL v1.2 corpus.
Use \code{OSSL_VisNIR} for any diffuse-reflectance VisNIR instrument
(350-2500 nm) and \code{OSSL_MIR} for any FTIR/ATR/DRIFTS MIR instrument
(600-4000 cm-1).
}
\keyword{datasets}
