% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_models.R
\name{models_available}
\alias{models_available}
\title{Check whether pre-trained models are available locally}
\usage{
models_available(
  family_id = c("OSSL_VisNIR", "OSSL_MIR"),
  model_dir = getOption("autoSpectra.model_dir", "models")
)
}
\arguments{
\item{family_id}{Family ID (\code{"OSSL_VisNIR"} or \code{"OSSL_MIR"})}

\item{model_dir}{Root model directory}
}
\value{
Named logical vector: \code{TRUE} if >= 5 .h5 files found
}
\description{
Check whether pre-trained models are available locally
}
