% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ossl_data.R
\name{ossl_join}
\alias{ossl_join}
\title{Join OSSL spectra with soil lab data for model training}
\usage{
ossl_join(spectra_mat, soillab_df, properties = ossl_l1_properties)
}
\arguments{
\item{spectra_mat}{Matrix from ossl_visnir_matrix() or ossl_mir_matrix()}

\item{soillab_df}{Data frame from ossl_soillab()}

\item{properties}{Properties to include; default all available}
}
\value{
Data frame with Soil_ID + spectral columns + property columns
}
\description{
Join OSSL spectra with soil lab data for model training
}
