% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{ossl_l1_labels}
\alias{ossl_l1_labels}
\title{Fancy display labels for OSSL L1 properties}
\format{
An object of class \code{character} of length 34.
}
\usage{
ossl_l1_labels
}
\description{
Fancy display labels for OSSL L1 properties
}
\keyword{datasets}
