% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{ossl_l1_properties}
\alias{ossl_l1_properties}
\title{All OSSL Level-1 harmonized soil property variable names}
\format{
An object of class \code{character} of length 34.
}
\usage{
ossl_l1_properties
}
\description{
All OSSL Level-1 harmonized soil property variable names
}
\keyword{datasets}
