% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ossl_data.R
\name{ossl_load_raw}
\alias{ossl_load_raw}
\title{Load a cached OSSL CSV (gzipped) as a data.table}
\usage{
ossl_load_raw(
  component = c("visnir", "mir", "soillab"),
  cache_dir = ossl_cache_dir()
)
}
\arguments{
\item{component}{"visnir", "mir", or "soillab"}

\item{cache_dir}{Cache directory}
}
\value{
data.table
}
\description{
Load a cached OSSL CSV (gzipped) as a data.table
}
