% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{ossl_mir_instruments}
\alias{ossl_mir_instruments}
\title{MIR instruments contributing to OSSL v1.2}
\format{
An object of class \code{character} of length 5.
}
\usage{
ossl_mir_instruments
}
\description{
MIR instruments contributing to OSSL v1.2
}
\keyword{datasets}
