% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{ossl_visnir_instruments}
\alias{ossl_visnir_instruments}
\title{VisNIR instruments contributing to OSSL v1.2}
\format{
An object of class \code{character} of length 5.
}
\usage{
ossl_visnir_instruments
}
\description{
VisNIR instruments contributing to OSSL v1.2
}
\keyword{datasets}
