% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{parse_sg}
\alias{parse_sg}
\title{Parse SG parameter string of the form "SG(m,p)", "SG(m,p,d)", etc.}
\usage{
parse_sg(st)
}
\arguments{
\item{st}{Character string like "SG(11,2,1)"}
}
\value{
Named list with m, p, d
}
\description{
Parse SG parameter string of the form "SG(m,p)", "SG(m,p,d)", etc.
}
