% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_scaler}
\alias{save_scaler}
\title{Save a z-score scaler to RDS and JSON}
\usage{
save_scaler(mean_y, sd_y, path_base)
}
\arguments{
\item{mean_y}{Mean of the target variable}

\item{sd_y}{Standard deviation of the target variable}

\item{path_base}{File path base (no extension)}
}
\description{
Save a z-score scaler to RDS and JSON
}
