% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sg_safe_params}
\alias{sg_safe_params}
\title{Compute safe SG parameters given available number of columns}
\usage{
sg_safe_params(m, p, ncols)
}
\arguments{
\item{m}{Half-window size (prospectr convention)}

\item{p}{Polynomial order}

\item{ncols}{Number of spectral columns}
}
\value{
Adjusted list(m, p), or NULL if impossible
}
\description{
Compute safe SG parameters given available number of columns
}
