% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vneed}
\alias{vneed}
\title{Safe shiny::validate wrapper}
\usage{
vneed(condition, message = "Validation failed.")
}
\arguments{
\item{condition}{Logical condition to test}

\item{message}{Error message to show if condition is FALSE}
}
\description{
Safe shiny::validate wrapper
}
